requirements.txt
pymupdf>=1.24.0
pymupdf4llm>=0.3.0
langchain-text-splitters>=0.2.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
flashrank>=0.2.0
rank-bm25>=0.2.0
ollama>=0.3.0
openai>=1.30.0
python-dotenv>=1.0.0
rich>=13.0.0
pytest>=8.0.0
